generate_input.sh no longer expands symbolic links.
Some computing facilities store data that can be
accessed through symbolic links. This becomes an
issue if generate_input.sh is used to create the
input file for the pipeline, which expands these
links before writing the paths to file. The data
paths in this file would then be in disagreement
with the symbolic link paths used to mount data.
The script no longer expands any symbolic links.